Feature

+++ src/main/java/com/HappyScrolls/service/ArticleService.java
@@ -3,6 +3,8 @@ package com.HappyScrolls.service;
 import com.HappyScrolls.dto.ArticleDTO;
 import com.HappyScrolls.entity.Article;
 import com.HappyScrolls.entity.Member;
+import com.HappyScrolls.exception.NoAuthorityExceoption;
+import com.HappyScrolls.exception.UserNotFoundException;
 import com.HappyScrolls.repository.ArticleRepository;
 import com.HappyScrolls.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class ArticleService {
@@ -49,7 +52,8 @@ public class ArticleService {
     }
 
     public ArticleDTO.Response articleRetrieve(Long id) {
-        Article article = articleRepository.findById(id).get();
+        Article article = articleRepository.findById(id).orElseThrow(()-> new NoSuchElementException(String.format("article[%s] 게시글을 찾을 수 없습니다", id))); //%s?
+
 
         return ArticleDTO.Response.builder()
                 .id(article.getId())
@@ -60,9 +64,12 @@ public class ArticleService {
 
     public ArticleDTO.Response articleEdit(Member member,ArticleDTO.Edit request) {
 
-        //유저 검증 로직
 
-        Article article = articleRepository.findById(request.getId()).get();
+        Article article = articleRepository.findById(request.getId()).orElseThrow(() -> new NoSuchElementException(String.format("article[%s] 게시글을 찾을 수 없습니다", request.getId()))); //%s?
+
+        if (!article.getMember().equals(member)) {
+            throw new NoAuthorityExceoption("수정 권한이 없습니다. 본인 소유의 글만 수정 가능합니다.");
+        }
 
         article.edit(request);
 
@@ -77,22 +84,30 @@ public class ArticleService {
 
     public void articleDelete(Member member,Long id) {
 
-        //유저 검증 로직
 
+        Article article = articleRepository.findById(id).orElseThrow(()-> new NoSuchElementException(String.format("article[%s] 게시글을 찾을 수 없습니다", id))); //%s?
 
-        Article article = articleRepository.findById(id).get();
+        if (!article.getMember().equals(member)) {
+            throw new NoAuthorityExceoption("삭제 권한이 없습니다. 본인 소유의 글만 삭제  가능합니다.");
+        }
 
         articleRepository.delete(article);
     }
 
-    public ArticleDTO.Response userArticleRetrieve(String email) {
-        Member findMember = memberRepository.findByEmail(email).get();
-        Article article = articleRepository.findByMember(findMember).get();
+    public List<ArticleDTO.Response>  userArticleRetrieve(String email) {
+        Member findMember = memberRepository.findByEmail(email).orElseThrow(()-> new UserNotFoundException(String.format("user[%s] 유저를  찾을 수 없습니다", email)));
+        List<Article> articles = articleRepository.findAllByMember(findMember);
 
-        return ArticleDTO.Response.builder()
-                .id(article.getId())
-                .title(article.getTitle())
-                .body(article.getBody())
-                .build();
+        List<ArticleDTO.Response> response = new ArrayList<>();
+
+        for (Article article : articles) {
+            response.add(ArticleDTO.Response.builder()
+                    .id(article.getId())
+                    .title(article.getTitle())
+                    .body(article.getBody())
+                    .build());
+        }
+
+        return response;
     }
 }
